factor program loading, parsing, and checking out of interpreter environment into checking environment
