+++ runtime/environment.go
@@ -23,13 +23,11 @@ import (
 
 	"github.com/onflow/cadence"
 	"github.com/onflow/cadence/activations"
-	"github.com/onflow/cadence/old_parser"
 
 	"github.com/onflow/cadence/ast"
 	"github.com/onflow/cadence/common"
 	"github.com/onflow/cadence/errors"
 	"github.com/onflow/cadence/interpreter"
-	"github.com/onflow/cadence/parser"
 	"github.com/onflow/cadence/sema"
 	"github.com/onflow/cadence/stdlib"
 )
@@ -84,7 +82,6 @@ type interpreterEnvironmentReconfigured struct {
 	runtimeInterface Interface
 	storage          *Storage
 	coverageReport   *CoverageReport
-	codesAndPrograms CodesAndPrograms
 }
 
 type interpreterEnvironment struct {
@@ -102,11 +99,10 @@ type interpreterEnvironment struct {
 	// by DeclareValue / interpreterBaseActivationFor
 	baseActivationsByLocation map[common.Location]*interpreter.VariableActivation
 
-	InterpreterConfig                     *interpreter.Config
-	CheckerConfig                         *sema.Config
+	InterpreterConfig *interpreter.Config
+
 	deployedContractConstructorInvocation *stdlib.DeployedContractConstructorInvocation
 	stackDepthLimiter                     *stackDepthLimiter
-	checkedImports                        importResolutionResults
 	compositeValueFunctionsHandlers       stdlib.CompositeValueFunctionsHandlers
 	config                                Config
 	deployedContracts                     map[Location]struct{}
@@ -139,7 +135,7 @@ func NewInterpreterEnvironment(config Config) *interpreterEnvironment {
 		stackDepthLimiter:     newStackDepthLimiter(config.StackDepthLimit),
 	}
 	env.InterpreterConfig = env.NewInterpreterConfig()
-	env.CheckerConfig = env.newCheckerConfig()
+
 	env.compositeValueFunctionsHandlers = stdlib.DefaultStandardLibraryCompositeValueFunctionHandlers(env)
 	return env
 }
@@ -177,18 +173,6 @@ func (e *interpreterEnvironment) NewInterpreterConfig() *interpreter.Config {
 	}
 }
 
-func (e *interpreterEnvironment) newCheckerConfig() *sema.Config {
-	return &sema.Config{
-		AccessCheckMode:                  sema.AccessCheckModeStrict,
-		BaseValueActivationHandler:       e.checkingEnvironment.getBaseValueActivation,
-		BaseTypeActivationHandler:        e.checkingEnvironment.getBaseTypeActivation,
-		ValidTopLevelDeclarationsHandler: validTopLevelDeclarations,
-		LocationHandler:                  e.ResolveLocation,
-		ImportHandler:                    e.resolveImport,
-		CheckHandler:                     newCheckHandler(&e.runtimeInterface),
-	}
-}
-
 func NewBaseInterpreterEnvironment(config Config) *interpreterEnvironment {
 	env := NewInterpreterEnvironment(config)
 	for _, valueDeclaration := range stdlib.DefaultStandardLibraryValues(env) {
@@ -212,11 +196,15 @@ func (e *interpreterEnvironment) Configure(
 	coverageReport *CoverageReport,
 ) {
 	e.runtimeInterface = runtimeInterface
-	e.codesAndPrograms = codesAndPrograms
 	e.storage = storage
 	e.InterpreterConfig.Storage = storage
 	e.coverageReport = coverageReport
 	e.stackDepthLimiter.depth = 0
+
+	e.checkingEnvironment.configure(
+		runtimeInterface,
+		codesAndPrograms,
+	)
 
 	configureVersionedFeatures(runtimeInterface)
 }
@@ -392,7 +380,7 @@ func (e *interpreterEnvironment) IsContractBeingAdded(location common.AddressLoc
 }
 
 func (e *interpreterEnvironment) TemporarilyRecordCode(location common.AddressLocation, code []byte) {
-	e.codesAndPrograms.setCode(location, code)
+	e.checkingEnvironment.temporarilyRecordCode(location, code)
 }
 
 func (e *interpreterEnvironment) ParseAndCheckProgram(
@@ -403,194 +391,7 @@ func (e *interpreterEnvironment) ParseAndCheckProgram(
 	*interpreter.Program,
 	error,
 ) {
-	return e.getProgram(
-		location,
-		func() ([]byte, error) {
-			return code, nil
-		},
-		getAndSetProgram,
-		importResolutionResults{
-			// Current program is already in check.
-			// So mark it also as 'already seen'.
-			location: true,
-		},
-	)
-}
-
-// parseAndCheckProgramWithRecovery parses and checks the given program.
-// It first attempts to parse and checks the program as usual.
-// If parsing or checking fails, recovery is attempted.
-//
-// Recovery attempts to parse the contract with the old parser,
-// and if it succeeds, uses the program recovery handler
-// to produce an elaboration for the old program.
-func (e *interpreterEnvironment) parseAndCheckProgramWithRecovery(
-	code []byte,
-	location common.Location,
-	checkedImports importResolutionResults,
-) (
-	program *ast.Program,
-	elaboration *sema.Elaboration,
-	err error,
-) {
-	// Attempt to parse and check the program as usual
-	program, elaboration, err = e.parseAndCheckProgram(
-		code,
-		location,
-		checkedImports,
-	)
-	if err == nil {
-		return program, elaboration, nil
-	}
-
-	// If parsing or checking fails, attempt to recover
-
-	recoveredProgram, recoveredElaboration := e.recoverProgram(
-		code,
-		location,
-		checkedImports,
-	)
-
-	// If recovery failed, return the original error
-	if recoveredProgram == nil || recoveredElaboration == nil {
-		return program, elaboration, err
-	}
-
-	recoveredElaboration.IsRecovered = true
-
-	// If recovery succeeded, return the recovered program and elaboration
-	return recoveredProgram, recoveredElaboration, nil
-}
-
-// parseAndCheckProgram parses and checks the given program.
-func (e *interpreterEnvironment) parseAndCheckProgram(
-	code []byte,
-	location common.Location,
-	checkedImports importResolutionResults,
-) (
-	program *ast.Program,
-	elaboration *sema.Elaboration,
-	err error,
-) {
-	wrapParsingCheckingError := func(err error) error {
-		switch err.(type) {
-		// Wrap only parsing and checking errors.
-		case *sema.CheckerError, parser.Error:
-			return &ParsingCheckingError{
-				Err:      err,
-				Location: location,
-			}
-		default:
-			return err
-		}
-	}
-
-	// Parse
-
-	reportMetric(
-		func() {
-			program, err = parser.ParseProgram(e, code, parser.Config{})
-		},
-		e.runtimeInterface,
-		func(metrics Metrics, duration time.Duration) {
-			metrics.ProgramParsed(location, duration)
-		},
-	)
-	if err != nil {
-		return nil, nil, wrapParsingCheckingError(err)
-	}
-
-	// Check
-
-	elaboration, err = e.check(location, program, checkedImports)
-	if err != nil {
-		return program, nil, wrapParsingCheckingError(err)
-	}
-
-	return program, elaboration, nil
-}
-
-// recoverProgram parses and checks the given program with the old parser,
-// and recovers the elaboration from the old program.
-func (e *interpreterEnvironment) recoverProgram(
-	oldCode []byte,
-	location common.Location,
-	checkedImports importResolutionResults,
-) (
-	program *ast.Program,
-	elaboration *sema.Elaboration,
-) {
-	// Parse
-
-	var err error
-	reportMetric(
-		func() {
-			program, err = old_parser.ParseProgram(e, oldCode, old_parser.Config{})
-		},
-		e.runtimeInterface,
-		func(metrics Metrics, duration time.Duration) {
-			metrics.ProgramParsed(location, duration)
-		},
-	)
-	if err != nil {
-		return nil, nil
-	}
-
-	// Recover elaboration from the old program
-
-	var newCode []byte
-	errors.WrapPanic(func() {
-		newCode, err = e.runtimeInterface.RecoverProgram(program, location)
-	})
-	if err != nil || newCode == nil {
-		return nil, nil
-	}
-
-	// Parse and check the recovered program
-
-	program, err = parser.ParseProgram(e, newCode, parser.Config{})
-	if err != nil {
-		return nil, nil
-	}
-
-	elaboration, err = e.check(location, program, checkedImports)
-	if err != nil || elaboration == nil {
-		return nil, nil
-	}
-
-	e.codesAndPrograms.setCode(location, newCode)
-
-	return program, elaboration
-}
-
-func (e *interpreterEnvironment) check(
-	location common.Location,
-	program *ast.Program,
-	checkedImports importResolutionResults,
-) (
-	elaboration *sema.Elaboration,
-	err error,
-) {
-	e.checkedImports = checkedImports
-
-	checker, err := sema.NewChecker(
-		program,
-		location,
-		e,
-		e.CheckerConfig,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	elaboration = checker.Elaboration
-
-	err = checker.Check()
-	if err != nil {
-		return nil, err
-	}
-
-	return elaboration, nil
+	return e.checkingEnvironment.ParseAndCheckProgram(code, location, getAndSetProgram)
 }
 
 func (e *interpreterEnvironment) ResolveLocation(
@@ -600,125 +401,7 @@ func (e *interpreterEnvironment) ResolveLocation(
 	res []ResolvedLocation,
 	err error,
 ) {
-	return ResolveLocationWithInterface(
-		e.runtimeInterface,
-		identifiers,
-		location,
-	)
-}
-
-func (e *interpreterEnvironment) resolveImport(
-	_ *sema.Checker,
-	importedLocation common.Location,
-	importRange ast.Range,
-) (sema.Import, error) {
-
-	// Check for cyclic imports
-	if e.checkedImports[importedLocation] {
-		return nil, &sema.CyclicImportsError{
-			Location: importedLocation,
-			Range:    importRange,
-		}
-	} else {
-		e.checkedImports[importedLocation] = true
-		defer delete(e.checkedImports, importedLocation)
-	}
-
-	const getAndSetProgram = true
-	program, err := e.GetProgram(
-		importedLocation,
-		getAndSetProgram,
-		e.checkedImports,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return sema.ElaborationImport{
-		Elaboration: program.Elaboration,
-	}, nil
-}
-
-func (e *interpreterEnvironment) GetProgram(
-	location Location,
-	storeProgram bool,
-	checkedImports importResolutionResults,
-) (
-	*interpreter.Program,
-	error,
-) {
-	return e.getProgram(
-		location,
-		func() ([]byte, error) {
-			return getLocationCodeFromInterface(e.runtimeInterface, location)
-		},
-		storeProgram,
-		checkedImports,
-	)
-}
-
-// getProgram returns the existing program at the given location, if available.
-// If it is not available, it loads the code, and then parses and checks it.
-func (e *interpreterEnvironment) getProgram(
-	location Location,
-	getCode func() ([]byte, error),
-	getAndSetProgram bool,
-	checkedImports importResolutionResults,
-) (
-	program *interpreter.Program,
-	err error,
-) {
-	load := func() (*interpreter.Program, error) {
-		code, err := getCode()
-		if err != nil {
-			return nil, err
-		}
-
-		e.codesAndPrograms.setCode(location, code)
-
-		parsedProgram, elaboration, err := e.parseAndCheckProgramWithRecovery(
-			code,
-			location,
-			checkedImports,
-		)
-		if parsedProgram != nil {
-			e.codesAndPrograms.setProgram(location, parsedProgram)
-		}
-		if err != nil {
-			return nil, err
-		}
-
-		return &interpreter.Program{
-			Program:     parsedProgram,
-			Elaboration: elaboration,
-		}, nil
-	}
-
-	if !getAndSetProgram {
-		return load()
-	}
-
-	errors.WrapPanic(func() {
-		program, err = e.runtimeInterface.GetOrLoadProgram(location, func() (program *interpreter.Program, err error) {
-			// Loading is done by Cadence.
-			// If it panics with a user error, e.g. when parsing fails due to a memory metering error,
-			// then do not treat it as an external error (the load callback is called by the embedder)
-			panicErr := UserPanicToError(func() {
-				program, err = load()
-			})
-			if panicErr != nil {
-				return nil, panicErr
-			}
-
-			if err != nil {
-				err = interpreter.WrappedExternalError(err)
-			}
-
-			return
-		})
-	})
-
-	return
+	return e.checkingEnvironment.resolveLocation(identifiers, location)
 }
 
 func (e *interpreterEnvironment) newInterpreter(
@@ -864,7 +547,7 @@ func (e *interpreterEnvironment) newImportLocationHandler() interpreter.ImportLo
 	return func(inter *interpreter.Interpreter, location common.Location) interpreter.Import {
 
 		const getAndSetProgram = true
-		program, err := e.GetProgram(
+		program, err := e.checkingEnvironment.GetProgram(
 			location,
 			getAndSetProgram,
 			importResolutionResults{},
